requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2>=2.8.6<2.9